third_party/blink/web_tests/external/wpt/dom/parts/basic-dom-part-declarative-brace-syntax-innerhtml.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Basic NodePart parsing
  assert_equals: There should be one NodePart expected 1 but got 0
Harness: the test ran to completion.
